Verifica que mkgmap y splitter hayan sido descargados antes de actualizar.

+++ bin/build_mkgmap.sh
@@ -7,14 +7,14 @@
 # * osmfilter
 # * pbftoosm
 #
-# (C) 2012 - 2022 Martin Andres Gomez Gimenez <mggimenez@nis.com.ar>
+# (C) 2012 - 2025 Martin Andres Gomez Gimenez <mggimenez@nis.com.ar>
 # Distributed under the terms of the GNU General Public License v3
 #
 
 
 WORKDIR=`pwd`
 MKGMAP_DOWNLOAD_URL="http://www.mkgmap.org.uk/download"
-MKGMAP_VERSION="r4922"
+MKGMAP_VERSION="r4923"
 SPLITTER_VERSION="r654"
 
 
@@ -28,20 +28,24 @@ PBFTOOSM="${WORKDIR}/bin/pbftoosm"
 
 
 
-rm -rf ${WORKDIR}/mkgmap*
-rm -rf ${WORKDIR}/splitter*
-
 wget -c ${MKGMAP_DOWNLOAD_URL}/mkgmap-${MKGMAP_VERSION}.zip
 wget -c ${MKGMAP_DOWNLOAD_URL}/splitter-${SPLITTER_VERSION}.zip
 
-unzip mkgmap-${MKGMAP_VERSION}.zip
-unzip splitter-${SPLITTER_VERSION}.zip
+# Verifica que mkgmap-${MKGMAP_VERSION}.zip fue descargado para actualizar.
+if [ -e mkgmap-${MKGMAP_VERSION}.zip ]; then
+  rm -rf ${WORKDIR}/mkgmap*
+  unzip mkgmap-${MKGMAP_VERSION}.zip
+  rm -f mkgmap-${MKGMAP_VERSION}.zip
+  ln --symbolic mkgmap-${MKGMAP_VERSION} mkgmap
+fi
 
-rm -f mkgmap-${MKGMAP_VERSION}.zip
-rm -f splitter-${SPLITTER_VERSION}.zip
-
-ln --symbolic mkgmap-${MKGMAP_VERSION} mkgmap
-ln --symbolic splitter-${SPLITTER_VERSION} splitter
+# Verifica que splitter-${SPLITTER_VERSION}.zip fue descargado para actualizar.
+if [ -e splitter-${SPLITTER_VERSION}.zip ]; then
+  rm -rf ${WORKDIR}/splitter*
+  unzip splitter-${SPLITTER_VERSION}.zip
+  rm -f splitter-${SPLITTER_VERSION}.zip
+  ln --symbolic splitter-${SPLITTER_VERSION} splitter
+fi
 
 wget -O - ${OSMCONVERT_URL} | cc -x c - -lz -O3 -o  ${OSMCONVERT} &>/dev/null
 wget -O -  ${OSMFILTER_URL} | cc -x c - -O3 -o ${OSMFILTER} &>/dev/null
